--- src/Entities/Comment.ts
@@ -1,68 +1,0 @@
-import { Entity, PrimaryGeneratedColumn, Column, ManyToOne } from "typeorm";
-import { ObjectType, Field, ID, InputType } from "type-graphql";
-import { User } from "./User";
-import { PointOfInterest } from "./PointOfInterest";
-
-@Entity()
-@ObjectType()
-export class Comment {
-  @PrimaryGeneratedColumn()
-  @Field(() => ID)
-  id: number;
-
-  @Column()
-  @Field()
-  comment: string;
-
-  @Column()
-  @Field()
-  note: number;
-
-  @Column()
-  @Field()
-  created_at: Date;
-
-  @Column({ nullable: true })
-  @Field({ nullable: true })
-  updated_at: Date;
-
-  @Column()
-  @Field()
-  created_by: User;
-
-  @Column({ nullable: true })
-  @Field({ nullable: true })
-  updated_by: User;
-
-  @ManyToOne(() => User, (user) => user.comments)
-  @Field(() => User)
-  user: User;
-
-  @ManyToOne(
-    () => PointOfInterest,
-    (pointOfInterest) => pointOfInterest.comments
-  )
-  @Field(() => [PointOfInterest])
-  pointOfInterest: PointOfInterest;
-}
-
-@InputType()
-export class CommentInput {
-  @Field()
-  comment: string;
-
-  @Field()
-  note: number;
-
-  @Field()
-  created_at: Date;
-
-  @Field({ nullable: true })
-  updated_at: Date;
-
-  @Field()
-  created_by: Date;
-
-  @Field({ nullable: true })
-  updated_by: Date;
-}
