+++ src/models/User.ts
@@ -1,0 +1,18 @@
+import { Entity, PrimaryGeneratedColumn, Column } from "typeorm";
+import { ObjectType, Field, ID } from "type-graphql";
+
+@Entity()
+@ObjectType()
+export class User {
+  @PrimaryGeneratedColumn()
+  @Field(() => ID, { nullable: true })
+  id: number;
+
+  @Column({ nullable: true })
+  @Field({ nullable: true })
+  email: string;
+
+  @Column({ nullable: true })
+  @Field({ nullable: true })
+  password: string;
+}
